Made multi-file and multi-patterns-file code work
Can now read multiple paths.
Can now read multiple patterns from a file. In this case it assumes the start of a pattern if there is a non-space character at the beginning of the line, adding all lines beginning with space characters to this line to create a pattern string.